% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhhrvcategoryTypeTable.R
\name{buttoncategoryTypeTableUI_left}
\alias{buttoncategoryTypeTableUI_left}
\title{操作区域}
\usage{
buttoncategoryTypeTableUI_left()
}
\value{
返回值
}
\description{
操作区域
}
\examples{
buttonUI_left()
}
