% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhhrvcategoryTypeTable.R
\name{buttoncategoryTypeTableUI_right}
\alias{buttoncategoryTypeTableUI_right}
\title{预览区域}
\usage{
buttoncategoryTypeTableUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
