% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhhrvtploutputvoucher.R
\name{buttontploutputvoucherUI_right}
\alias{buttontploutputvoucherUI_right}
\title{预览区域}
\usage{
buttontploutputvoucherUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
